// src/types/Houses.ts
type URL = string
export interface HouseIceAndFire {
  url: URL
  name: string
  region: string
  coatOfArms: string
  words: string
  titles: any[]
  seats: string[]
  currentLord: URL
  heir: URL
  overlord: URL
  founded: string
  founder: string
  diedOut: string
  ancestralWeapons: string[]
  cadetBranches: URL[]
  swornMembers: URL[]
}